+++ cogs/gatekeeper/__init__.py
@@ -1,6 +1,0 @@
-from .gatekeeper import Gatekeeper
-
-
-def setup(bot):
-    """Load cog"""
-    bot.add_cog(Gatekeeper(bot))